Refactoring + todo demo
Refactoring (alot)!
new:
- Todo list demo (inspired by http://javascriptmvc.com/todo)
- .FindElement() method
- .GetViewId() method
- .KeyCheck() "static" method
- .List() class/method with a more human way of handling array's in
javascript. it extends the array (currently) with 2+ extra methods,
which makes it more semantic when adding and removing elements using
.Add() and .Remove() methods respectively.
- Added events: .mouseenter, .mouseleave
- Added comments
- Renamed .ExecuteController() to .RunCtr()
- "Chainablitiy" - the mutator methods now return the object itself, so
its possible to chain methods (like jQuery).
- Added .init() method, which can be implemented with the data. It
executes 1 ms after instantiation of the object.

Fixed a bug when setting values: Changed .text() to .html() ! line 843
Fixed a bug when attaching events to the buttons which correspond witht
the methods in the controller! Line 200.