Refactoring
Added:
* datasrc property to the settings, as it should be possible to create
a ModelView without any data, but solely relies on receiving updated
values from another view!

Changed:
* .Set() to update databound elements with datasrc if specified,
otherwise with viewId.
* .FindElement() renamed to .Find()

Fixed:
* When setting the values on .init() the databound items should
ofcourse get updated even if the model data is same as in the view.
moved '$object.SetDataboundDomVal(viewId, key, newVal);' out of
if(oldVal === newVal) on line 598.
* When cloning, 'withDataAndEvents' would be undefined if not
specified. Now it will be 'false' if not specified, which was intented
behaviour.