Fixed bugs with .Get & .Set
The tests failed when a property was not set explicitly in the Model &
View. This has been fixed, so that if isMirror is TRUE & the property
isn’t in the Model, it will be added. Before this fix it would not be
possible to use .Get(‘properyName’) unless the property was in the
Model or View. Also .Set() method now uses viewId if datasrc is
undefined / null. Fixed span element to used datafld attribute instead
of ID. Custom html elements are now supported regarding auto-binding of
methods in the Controller with the View. NOTE: ‘isEvent’-class must be
added to the element and have a name attribute.