Bug fix + added functionality
Fixed fundamental flaw with the library, where 2 or more instances of
Simple MVC (the $().ModelView(…)) would interact with each other. This
was fixed in SetDataboundDomVal().

FillSelect() method added, which requires a property name (type of
array), text and value, to fill an select element with option values.
This makes working with select much easier.