Added new functionality
Added Sort() and Reverse() methods to MVC.List.
Added New() method which creates a new object from a view.
Added sorting functionality and Comparator to $object.
Select (lists) are sorted ascending on default. Should be an option to
choose if ascending or descending or if any sorting is used at all.